[TM-3425] restructure files
