+++ src/pages/site/[uuid]/tabs/SitePolygonsTab.tsx
@@ -2,16 +2,14 @@ import { FC } from "react";
 
 import { SiteFullDto } from "@/generated/v3/entityService/entityServiceSchemas";
 
-import SitePolygonsReviewTab from "./SitePolygonsReviewTab";
+import SitePolygonsWorkspace from "../sitePolygonReview/SitePolygonsWorkspace";
 
 interface SitePolygonsTabProps {
   site: SiteFullDto;
 }
 
-export type { PolygonTableRow } from "../components/PolygonTableRow";
-
 const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => (
-  <SitePolygonsReviewTab site={site} variant="champions" />
+  <SitePolygonsWorkspace site={site} variant="champions" />
 );
 
 export default SitePolygonsTab;
